A model class for conda environment parameters.
