+++ src/main/java/fiji/plugin/trackmate/helper/ui/components/ArrayRangeSweepPanel.java
@@ -8,12 +8,12 @@
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public
  * License along with this program.  If not, see
  * <http://www.gnu.org/licenses/gpl-3.0.html>.
@@ -45,8 +45,8 @@ import javax.swing.border.EmptyBorder;
 
 import fiji.plugin.trackmate.gui.Fonts;
 import fiji.plugin.trackmate.gui.Icons;
+import fiji.plugin.trackmate.helper.model.parameter.AbstractArrayParamSweepModel.RangeType;
 import fiji.plugin.trackmate.helper.model.parameter.ArrayParamSweepModel;
-import fiji.plugin.trackmate.helper.model.parameter.ArrayParamSweepModel.RangeType;
 
 public class ArrayRangeSweepPanel< T > extends JPanel
 {
